도메인 클래스에 대한 중복 제거 및 리팩토링
중복되는 코드를 제거하기 위해서 부모클래스를 생성하여 공통 기능을 넣음
createdAt, modifiedDate를 데이터가 변경될 때 마다 자동으로 업데이트 퇴도록 수정함
