+++ src/main/java/com/codessquad/qna/domain/Question.java
@@ -3,16 +3,10 @@ package com.codessquad.qna.domain;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import javax.persistence.*;
-import java.time.LocalDateTime;
 import java.util.List;
 
 @Entity
-public class Question {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+public class Question extends AbstractEntity {
     @ManyToOne
     @JoinColumn(foreignKey = @ForeignKey(name = "fk_question_writer"))
     private User writer;
@@ -22,9 +16,6 @@ public class Question {
 
     @Column(nullable = false)
     private String contents;
-
-    @Column(nullable = false)
-    private LocalDateTime createdAt;
 
     @OrderBy("id DESC")
     @OneToMany(mappedBy = "question")
@@ -36,20 +27,8 @@ public class Question {
     @JsonProperty
     private Integer countOfAnswer = 0;
 
-    public Question() {
-        markCreatedTime();
-    }
-
     public List<Answer> getAnswers() {
         return answers;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public LocalDateTime getCreated() {
-        return createdAt;
     }
 
     public User getWriter() {
@@ -78,10 +57,6 @@ public class Question {
 
     public int getCountOfAnswer() {
         return this.countOfAnswer;
-    }
-
-    private void markCreatedTime() {
-        this.createdAt = LocalDateTime.now();
     }
 
     public void update(Question question) {
@@ -128,11 +103,10 @@ public class Question {
     @Override
     public String toString() {
         return "Question{" +
+                super.toString() +
                 "writer='" + writer + '\'' +
                 ", title='" + title + '\'' +
                 ", contents='" + contents + '\'' +
-                ", createdAt='" + createdAt + '\'' +
-                ", id=" + id +
                 '}';
     }
 }
